removed dublicated folder and moved DependecyControllerTest to correct testing file
