+++ src/test/java/com/inso/MinecraftProject/DependencyControllerTest.java
@@ -1,9 +1,11 @@
-package com.inso.MinecraftProject.service;
+package com.inso.MinecraftProject;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
+
+import com.inso.MinecraftProject.service.DependencyController;
 
 import static org.junit.jupiter.api.Assertions.assertTrue; // cambio: para validar que timeout sí retrasó la respuesta
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
